Update: add server url Swagger and change administrativeFee data type to int

+++ src/main/java/com/project/zipkok/model/RealEstate.java
@@ -41,7 +41,7 @@ public class RealEstate {
     private long price;
 
     @Column(name = "administrative_fee")
-    private float administrativeFee;
+    private int administrativeFee;
 
     @Column(name = "detail")
     private String detail;
